handle api error response

+++ rss_reader/lib/api.py
@@ -6,32 +6,35 @@ from rss_reader.lib.model import db, Feed, Entry
 from rss_reader.parser import parse
 
 
-bp = Blueprint("api", __name__, url_prefix="/api")
+api = Blueprint("api", __name__, url_prefix="/api")
 
 
-def login_required(func):
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        if not current_user.is_authenticated:
-            abort(401)
-        return func(*args, **kwargs)
-    return wrapper
+def api_response(login_required: bool = True):
+    def decorator(func):
+        @wraps(func)
+        def wrapper(*args, **kwargs):
+            if login_required and not current_user.is_authenticated:
+                abort(401)
+            data, status = func(*args, **kwargs)
+            return jsonify({"data": data, "status": status})
+        return wrapper
+    return decorator
 
 
-@bp.route("/feeds/", methods=["GET"])
-@login_required
+@api.route("/feeds/", methods=["GET"])
+@api_response(login_required=False)
 def list_feeds():
     feeds = Feed.query.filter(Feed.user_id == current_user.id).all()
-    return jsonify([feed.to_json() for feed in feeds])
+    return ([feed.to_json() for feed in feeds], "ok")
 
 
-@bp.route("/feeds/", methods=["POST"])
-@login_required
+@api.route("/feeds/", methods=["POST"])
+@api_response(login_required=True)
 def add_feed():
     try:
         parser = parse(request.form.get("uri"))
     except Exception:
-        abort(400)
+        return (None, "parser_error")
 
     feed = Feed(
         user_id=current_user.id,
@@ -51,40 +54,46 @@ def add_feed():
     db.session.flush()
     db.session.commit()
 
-    return jsonify(feed.to_json())
+    return (feed.to_json(), "ok")
 
 
-@bp.route("/feeds/<int:feed_id>/", methods=["GET"])
-@login_required
+@api.route("/feeds/<int:feed_id>/", methods=["GET"])
+@api_response(login_required=True)
 def get_feed(feed_id: int):
-    feed = Feed.query.get_or_404(feed_id)
+    feed = Feed.query.get(feed_id)
+    if not feed:
+        return (None, "not_found")
     if feed.user_id != current_user.id:
-        abort(403)
-    return jsonify(feed.to_json())
+        return (None, "permission_denied")
+    return (feed.to_json(), "ok")
 
 
-@bp.route("/feeds/<int:feed_id>/", methods=["DELETE"])
-@login_required
+@api.route("/feeds/<int:feed_id>/", methods=["DELETE"])
+@api_response(login_required=True)
 def delete_feed(feed_id: int):
-    feed = Feed.query.get_or_404(feed_id)
+    feed = Feed.query.get(feed_id)
+    if not feed:
+        return (None, "not_found")
     if feed.user_id != current_user.id:
-        abort(403)
+        return (None, "permission_denied")
     db.session.delete(feed)
     db.session.commit()
-    return jsonify({})
+    return (None, "ok")
 
 
-@bp.route("/feeds/<int:feed_id>/entries/", methods=["GET"])
-@login_required
+@api.route("/feeds/<int:feed_id>/entries/", methods=["GET"])
+@api_response(login_required=True)
 def list_feed_entries(feed_id: int):
-    feed = Feed.query.get_or_404(feed_id)
+    feed = Feed.query.get(feed_id)
+    if not feed:
+        return (None, "not_found")
     if feed.user_id != current_user.id:
-        abort(403)
-    return jsonify([entry.to_json() for entry in feed.entries])
+        return (None, "permission_denied")
+    return ([entry.to_json() for entry in feed.entries], "ok")
 
 
-@bp.route("/feeds/entries/", methods=["GET"])
-@login_required
+@api.route("/feeds/entries/", methods=["GET"])
+@api_response(login_required=True)
 def list_all_entries():
     entries = Entry.query.filter(Entry.user_id == current_user.id).all()
-    return jsonify([entry.to_json() for entry in entries])
+    return ([entry.to_json() for entry in entries], "ok")
